+++ Monstre.pde
@@ -1,10 +1,10 @@
 class Monstre extends LeftRunner {
 
-  Monstre(int x, int y, int s) {
+  Monstre(int x, int y, float s) {
     super(x, y, s);
   }
 
-  Monstre(int x, int y, int s, int mY) {
+  Monstre(int x, int y, float s, int mY) {
     super(x, y, s, mY);
   }
